+++ src/manifests/applications/design.ignaz.polarisos.test.manifest.js
@@ -1,15 +1,14 @@
 import ClockIcon from "../../components/icons/ClockIcon.vue";
 import { useWindowManager } from "../../composables/windowManager";
 import { defineAsyncComponent, markRaw, watch } from "vue";
-const { register, processIDsWithOpenWindows } = useWindowManager();
+const { createWindow, processIDsWithOpenWindows } = useWindowManager();
 
 export default {
   name: "test application from manifest!",
   icon: "cd",
   reEntry: async (processID) => {
-    const parentWindowID = register(
+    const parentWindow = createWindow(
       processID,
-      "anotherWindow",
       defineAsyncComponent(
         () => import("../../components/applications/AboutApplication.vue"),
       ),
@@ -17,9 +16,8 @@ export default {
     );
 
     // Child window!
-    register(
+    createWindow(
       processID,
-      "window",
       defineAsyncComponent(
         () => import("../../components/applications/ClockApplication.vue"),
       ),
@@ -30,12 +28,11 @@ export default {
         },
         // alwaysOnTop: true,
       },
-      parentWindowID,
+      parentWindow.windowID,
     );
 
-    register(
+    createWindow(
       processID,
-      "thirdWindow",
       defineAsyncComponent(
         () => import("../../components/applications/AboutApplication.vue"),
       ),